added test case, now supports portworx storage by default

+++ .github/scripts/validate-deploy.sh
@@ -63,9 +63,40 @@ if [[ $count -eq 20 ]]; then
   exit 1
 fi
 
-kubectl get Ibmcpd ibmcpd-cr
-kubectl get ZenService lite-cr 
-#kubectl get ZenService lite-cr -o jsonpath="{.status.url}{'\n'}"
+STATUS=$(kubectl get Ibmcpd ibmcpd-cr -n "${NAMESPACE}" -o jsonpath="{.status.controlPlaneStatus}")
+count=0
+until [[ $STATUS == "Completed" ]] || [[ $count -eq 20 ]]; do
+  echo "ibmcpd/ibmcpd-cr status: ${STATUS}"
+  STATUS=$(kubectl get Ibmcpd ibmcpd-cr -n "${NAMESPACE}" -o jsonpath="{.status.controlPlaneStatus}")
+  count=$((count + 1))
+  sleep 15
+done
+
+if [[ $count -eq 20 ]]; then
+  echo "Timed out waiting for ibmcpd/ibmcpd-cr to achieve Completed status"
+  kubectl get ibmcpd ibmcpd-cr -n "${NAMESPACE}" -o yaml
+  exit 1
+fi
+
+
+
+STATUS=$(kubectl get ZenService lite-cr -n "${NAMESPACE}" -o jsonpath="{.status.zenStatus}")
+count=0
+until [[ $STATUS == "Completed" ]] || [[ $count -eq 20 ]]; do
+  echo "ZenService/lite-cr status: ${STATUS}"
+  STATUS=$(kubectl get ZenService lite-cr -n "${NAMESPACE}" -o jsonpath="{.status.zenStatus}")
+  count=$((count + 1))
+  sleep 15
+done
+
+if [[ $count -eq 20 ]]; then
+  echo "Timed out waiting for ZenService/lite-cr to achieve Completed status"
+  kubectl get get ZenService lite-cr -n "${NAMESPACE}" -o yaml
+  exit 1
+fi
+
+kubectl get get ZenService lite-cr -n "${NAMESPACE}" -o yaml
+
 
 cd ..
 rm -rf .testrepo
